+++ lithops/serverless/backends/cloudrun/config.py
@@ -53,3 +53,7 @@ def load_config(config_data):
     if 'workers' not in config_data['lithops']:
         config_data['cloudrun']['workers'] = CONCURRENT_WORKERS_DEFAULT
         config_data['lithops']['workers'] = CONCURRENT_WORKERS_DEFAULT
+
+    if 'invoke_pool_threads' not in config_data['cloudrun']:
+        config_data['cloudrun']['invoke_pool_threads'] = config_data['lithops']['workers']
+    config_data['serverless']['invoke_pool_threads'] = config_data['cloudrun']['invoke_pool_threads']