+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 - [Config] Allow 'log_level' and 'log_format' keys in configuration
 - [Config] Allow 'log_stream' and 'log_filename' keys in configuration
+- [Config] Allow 'runtime' being configured at serverless backend level
+- [Config] Allow 'invoke_pool_threads' being configured at serverless backend level
 
 ### Changed
 - [Core] Renamed utils.setup_logger() method to utils.setup_lithops_logger()